==> scripts/2_prep/3_copy_files.ipynb <==
import pandas as pd
import os
import shutil
from tqdm import tqdm

#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/sampled_ds_bothA_350_split.csv')
#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/sampled_ds_ground_620.csv')
#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/sampled_ds_trail_180.csv')

#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/ground_split.csv')
#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/trail_split.csv')
#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/both_split.csv')

#df = pd.read_csv('/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/demo2.csv')
df = pd.read_csv('/Volumes/Cara_cam_traps/CV4E/data_cleaned/demo2/demo2_empties.csv')

df.head()

df['full_path'].values
df['Full Path'].values #for empties

for file in tqdm(df['full_path'].values):
    
    #file_source = '/Users/caraappel/Documents/CV4E/' + file
    #file_target = '/Users/caraappel/Documents/CV4E/upload_trail_thresh180/' + file
    #file_target = '/Volumes/Cara_cam_traps/CV4E/data_cleaned/both/' + file

    #for empties:
    file_source = file
    file_target = '/Volumes/Cara_cam_traps/CV4E/data_cleaned/demo2/data/' + file

    os.makedirs(os.path.dirname(file_target), exist_ok=True)

    shutil.copy(file_source, file_target)
    

  

## For empties: (messy but just move them later... ugh)
for file in tqdm(df['Full Path'].values):
    
    file_source = file
    file_target = '/Volumes/Cara_cam_traps/CV4E/data_cleaned/demo2/data/' + file

    os.makedirs(os.path.dirname(file_target), exist_ok=True)

    shutil.copy(file_source, file_target)
