==> scripts/2_prep/5_csv_to_coco.ipynb <==
# Convert CSV to COCO JSON file

import csv
import json
import os
from PIL import Image

# Set working directory
os.chdir('/Users/caraappel/Documents/CV4E/oregon_critters')
print(os.getcwd())

# Define function to read CSV and reformat as COCO JSON
def csv_to_coco(csv_file_path, output_json_path):
    coco_data = {
        "info": {
            "year": 2023,
            "version": "1.0",
            "description": "CSV to COCO conversion",
            "contributor": "",
            "url": "",
            "date_created": "2023-08-18"
        },
        "licenses": [],
        "categories": [],
        "images": [],
        "annotations": []
    }

    category_id_map = {
    "Bat": 0,
    "CanadaJay": 1,
    "Coyote": 2,
    "PileatedWoodpecker": 3,
    "Owl": 4,
    "Mink": 5,
    "RedFox": 6,
    "Marten": 7,
    "AmericanRobin": 8,
    "DouglasSquirrel": 9,
    "FlyingSquirrel": 10,
    "Grouse": 11,
    "Invertebrate": 12,
    "MountainBeaver": 13,
    "OtherBird": 14,
    "SmallMammal": 15,
    "Warbler": 16,
    "Chickadee": 17,
    "MountainQuail": 18,
    "StripedSkunk": 19,
    "Human": 20,
    "Dog": 21,
    "BlackBear": 22,
    "BlackTailedDeer": 23,
    "Bobcat": 24,
    "Cougar": 25,
    "DarkEyedJunco": 26,
    "GraySquirrel": 27,
    "Herp": 28,
    "NorthernFlicker": 29,
    "Opossum": 30,
    "Rabbit": 31,
    "Racoon": 32,
    "RooseveltElk": 33,
    "SpottedSkunk": 34,
    "StellersJay": 35,
    "Thrush": 36,
    "TownsendsChipmunk": 37,
    "VariedThrush": 38,
    "Weasel": 39,
    "Woodpecker": 40,
    "Woodrat": 41,
    "Wren": 42,
    "GroundSquirrel": 43,
    "GrayFox": 44
}


    # Populate the "categories" section
    categories = []
    for class_name, class_id in category_id_map.items():
        category_info = {
            "id": class_id,
            "name": class_name,
            "supercategory": "object"
        }
        categories.append(category_info)
    
    coco_data['categories'] = categories
    
    image_id_map = {}
    annotation_id = 1

    with open(csv_file_path, 'r') as csv_file:
        csv_reader = csv.DictReader(csv_file)
        for row in csv_reader:
            filename = row['full_path']
            class_name_from_csv = row['CName']
            desired_class_id = category_id_map.get(class_name_from_csv)

            if desired_class_id is not None:  # Skip rows with unknown classes
                xmin = float(row['X'])
                ymin = float(row['Y'])
                width = float(row['W'])
                height = float(row['H'])
                img_width = float(row['img_width'])
                img_height = float(row['img_height'])

                if filename in image_id_map:
                    image_id = image_id_map[filename]
                else:
                    image_id = len(coco_data['images']) + 1
                    image_id_map[filename] = image_id

                    image_info = {
                        "id": image_id,
                        "file_name": filename,
                        "width": img_width,
                        "height": img_height
                    }

                    coco_data['images'].append(image_info)

                annotation = {
                    "id": annotation_id,
                    "image_id": image_id,
                    "category_id": desired_class_id,
                    "bbox": [xmin, ymin, width, height],
                    "area": width * height,
                    "iscrowd": 0
                }
                coco_data['annotations'].append(annotation)
                annotation_id += 1

    with open(output_json_path, 'w') as output_json_file:
        json.dump(coco_data, output_json_file, indent=4)

# Read in CSV, convert, and output to COCO JSON

#csv_file_path = '/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/trail_split_dim.csv'
#output_json_path = '/Users/caraappel/Documents/CV4E/oregon_critters/coco_json/trail.json'
#csv_to_coco(csv_file_path, output_json_path)

# For demo dataset
csv_file_path = '/Users/caraappel/Documents/CV4E/oregon_critters/metadata_labels/demo2.csv'
output_json_path = '/Users/caraappel/Documents/CV4E/oregon_critters/coco_json/demo2.json'
csv_to_coco(csv_file_path, output_json_path)
