==> scripts/2_prep/6_visualize_boxes.ipynb <==
from PIL import Image, ImageDraw, ImageFont
import json
from IPython.display import display

# Specify directory where the images are located
image_directory = "/Users/caraappel/Documents/CV4E"

# Load COCO JSON file
coco_json_path = '/Users/caraappel/Documents/CV4E/oregon_critters/coco_json/both.json'
with open(coco_json_path, 'r') as json_file:
    coco_data = json.load(json_file)


# Define function to visualize annotations for a specific image
def visualize_annotations_for_image(image_info, annotations, categories):
    image_path = image_directory + "/" + image_info['file_name']
    image = Image.open(image_path)
    draw = ImageDraw.Draw(image)
 
    # Customize font and fill color for the label text
    #label_font = ImageFont.load_default()  
    label_font = ImageFont.truetype('/Users/caraappel/Documents/Inclusive_Sans/InclusiveSans-Regular.ttf', 36)
    label_fill = 'white'  
    
    for ann in annotations:
        
        # Draw the bounding box
        bbox = ann['bbox']
        x1, y1, width, height = bbox
        x2, y2 = x1 + width, y1 + height
        draw.rectangle([x1, y1, x2, y2], outline='red', width=5)
        
        category_id = ann['category_id']
        category_name = next((cat['name'] for cat in categories if cat['id'] == category_id), "Unknown")
        label_text = f"{category_name}"
        
        # Draw the textbox
        label_bbox = draw.textbbox((x1, y1), label_text, font=label_font)
        draw.rectangle(label_bbox, fill='black')
       
        # Add label text
        draw.text((label_bbox[0], label_bbox[1]), label_text, fill=label_fill, font=label_font)

    #display(image)  #to display the image directly in the notebook
    image.show()     #to open a new window with the image


# Choose an image ID to visualize
image_id_to_visualize = 780  

# Find the image information, annotations, and categories for the specified image ID
image_info = next((img for img in coco_data['images'] if img['id'] == image_id_to_visualize), None)
annotations = [ann for ann in coco_data['annotations'] if ann['image_id'] == image_id_to_visualize]
categories = coco_data['categories']

# Visualize the bounding boxes and labels for the specified image
if image_info and annotations:
    visualize_annotations_for_image(image_info, annotations, categories)
else:
    print("Image information or annotations not found.")
